fix: Legacy4J GLFW controller enable
Legacy4J previously was not being detected by the GLFW implementation as
a mod that was using GLFW controllers, thus leading it to use vanilla
keyboard+mouse emulation. This fixes that.

This is still kinda bugged on L4J settings menu but this only applies
when outside of a world. I can't figure out why L4J does this only on
their menus, I probably wont either, SDL doesn't have this.