SDL: Improve controllable mitigation
Now continously polls the directory where controllable is expected to
extract natives instead of waiting on inotify. This theoretically should
consume quite a bit of the CPU but it seems to not when testing.

This should be much more consistent than waiting on inotify, testing
seems to corroborate that. Anyone who's crashing can send a bug report.